+++ Source/screen.cpp
@@ -18,9 +18,9 @@ clsScreen::clsScreen() {
         else {height = Global::Cnfg.getvalues(cnfgScreenHeight);}
 
         //Set all the booleans to false
-        blnWindow = false;
-        blnRenderer = false;
-        blnSky = blnPlayer = blnMonster = blnWall = blnPole = blnCoin = blnErrortex = false;
+        blnloaded.blnWindow = false;
+        blnloaded.blnRenderer = false;
+        blnloaded.blnSky = blnloaded.blnPlayer = blnloaded.blnMonster = blnloaded.blnWall = blnloaded.blnPole = blnloaded.blnCoin = blnloaded.blnErrortex = false;
         bln_SDL_started = false;
 
         //Start SDL
@@ -54,7 +54,7 @@ clsScreen::clsScreen() {
             error();
             return;
         } else {
-            blnWindow = true;
+            blnloaded.blnWindow = true;
             if (Global::blnDebugMode) {printf("Window creation successful\n");}
         }
 
@@ -64,68 +64,66 @@ clsScreen::clsScreen() {
             error();
             return;
         } else {
-            blnRenderer = true;
+            blnloaded.blnRenderer = true;
             if (Global::blnDebugMode) {printf("Renderer creation successful\n");}
         }
 
-
-
-        errortex = loadERROR();
+        textures.errortex = loadERROR();
         if (bln_SDL_started == false) {return;}
         else {
-            blnErrortex = true;
+            blnloaded.blnErrortex = true;
             if (Global::blnDebugMode) {printf("Error loading successful\n");}
         }
 
         std::string path = DEFINED_DEFAULT_IMAGE_PATH;
         path += "sky.png";
-        sky = loadIMG(path);
-        blnSky = true;
+        textures.sky = loadIMG(path);
+        blnloaded.blnSky = true;
         if (Global::blnDebugMode) {printf("Sky loading successful\n");}
 
         path = DEFINED_DEFAULT_IMAGE_PATH;
         path += "player.png";
-        player = loadIMG(path);
-        blnPlayer = true;
+        textures.player = loadIMG(path);
+        blnloaded.blnPlayer = true;
         if (Global::blnDebugMode) {printf("Player loading successful\n");}
 
 
         path = DEFINED_DEFAULT_IMAGE_PATH;
         path += "wall.png";
-        wall = loadIMG(path);
-        blnWall = true;
+        textures.wall = loadIMG(path);
+        blnloaded.blnWall = true;
         if (Global::blnDebugMode) {printf("Wall loading successful\n");}
 
 
         path = DEFINED_DEFAULT_IMAGE_PATH;
         path += "coin.png";
-        coin = loadIMG(path);
-        blnCoin = true;
+        textures.coin = loadIMG(path);
+        blnloaded.blnCoin = true;
         if (Global::blnDebugMode) {printf("Coin loading successful\n");}
 
         path = DEFINED_DEFAULT_IMAGE_PATH;
         path += "pole.png";
-        pole = loadIMG(path);
-        blnPole = true;
+        textures.pole = loadIMG(path);
+        blnloaded.blnPole = true;
         if (Global::blnDebugMode) {printf("Pole loading successful\n");}
 
         path = DEFINED_DEFAULT_IMAGE_PATH;
         path += "monster.png";
-        monster = loadIMG(path);
-        blnMonster = true;
+        textures.monster = loadIMG(path);
+        blnloaded.blnMonster = true;
         if (Global::blnDebugMode) {printf("Monster loading successful\n");}
 
         MessageFont = TTF_OpenFont(DEFINED_MESSAGE_FONT,16); //Opens font and sets size
         if (MessageFont == nullptr) {
             printf("Font failed to load, messages will not appear.");
-            blnMessageFont = false;
+            blnloaded.blnMessageFont = false;
         } else {
             if(Global::blnDebugMode) {printf("Message font created\n");}
-            blnMessageFont = true;
+           blnloaded.blnMessageFont = true;
         }
 
-        Black = {0, 0, 0, 0}; //Make the color black for fonts
-        White = {255, 255, 255, 0}; //Make the color white for fonts
+        colors.Black = {0, 0, 0, 0}; //Make the color black for fonts
+        colors.White = {255, 255, 255, 0}; //Make the color white for fonts
 
         update();
     } //end if blnShowMap
@@ -155,7 +153,7 @@ void clsScreen::update(void) {
     SDL_RenderClear(ren);
 
     //copy sky to cover entire screen.
-    SDL_RenderCopy(ren,sky,NULL,NULL);
+    SDL_RenderCopy(ren,textures.sky,NULL,NULL);
 
     SDL_Rect dst;
 
@@ -168,37 +166,37 @@ void clsScreen::update(void) {
             //Query a texture to get its width and height
             //Since all the textures are the same we'll use the error one since the program would
             //not have gotten this far if that failed load.
-            SDL_QueryTexture(errortex,NULL,NULL, &dst.w, &dst.h);
+            SDL_QueryTexture(textures.errortex,NULL,NULL, &dst.w, &dst.h);
 
             switch(Global::Map.getMapCell(x,y)) {
             case tileSpace:
                 //SDL_RenderCopy(ren,sky,NULL, &dst);
                 break;
             case tileCoin:
-                SDL_RenderCopy(ren,coin,NULL, &dst);
+                SDL_RenderCopy(ren,textures.coin,NULL, &dst);
                 break;
             case tileMonster:
-                SDL_RenderCopy(ren, monster, NULL, &dst);
+                SDL_RenderCopy(ren, textures.monster, NULL, &dst);
                 break;
             case tilePlayer:
-                SDL_RenderCopy(ren, player, NULL, &dst);
+                SDL_RenderCopy(ren, textures.player, NULL, &dst);
                 break;
             case tilePole:
-                SDL_RenderCopy(ren, pole, NULL, &dst);
+                SDL_RenderCopy(ren, textures.pole, NULL, &dst);
                 break;
             case tileWall:
-                SDL_RenderCopy(ren, wall, NULL, &dst);
+                SDL_RenderCopy(ren, textures.wall, NULL, &dst);
                 break;
             default:
                 //Don't know what this is so display an error texture.
-                SDL_RenderCopy(ren, errortex, NULL, &dst);
+                SDL_RenderCopy(ren, textures.errortex, NULL, &dst);
                 break;
             } //end switch
         } //end for x
     } //end for y
 
     //Write messages only if Message font is loaded.
-    if (blnMessageFont) {
+    if (blnloaded.blnMessageFont) {
         writemessage();
     }
     //show renderer
@@ -207,67 +205,67 @@ void clsScreen::update(void) {
 }
 /**********************************************************************************************************************************************************************/
 void clsScreen::cleanup(void) {
-    if (blnPlayer) {
-        SDL_DestroyTexture(player);
-        blnPlayer = false;
+    if (blnloaded.blnPlayer) {
+        SDL_DestroyTexture(textures.player);
+        blnloaded.blnPlayer = false;
         if (Global::blnDebugMode) {printf("Player texture destroyed\n");}
     }
-	if (blnSky) {
-        SDL_DestroyTexture(sky);
-        blnSky = false;
+	if (blnloaded.blnSky) {
+        SDL_DestroyTexture(textures.sky);
+        blnloaded.blnSky = false;
         if (Global::blnDebugMode) {printf("Sky texture destroyed\n");}
     }
 
-    if (blnCoin) {
-        SDL_DestroyTexture(coin);
-        blnCoin = false;
+    if (blnloaded.blnCoin) {
+        SDL_DestroyTexture(textures.coin);
+        blnloaded.blnCoin = false;
         if (Global::blnDebugMode) {printf("Coin texture destroyed\n");}
     }
 
-     if (blnWall) {
-        SDL_DestroyTexture(wall);
-        blnWall = false;
+     if (blnloaded.blnWall) {
+        SDL_DestroyTexture(textures.wall);
+        blnloaded.blnWall = false;
         if (Global::blnDebugMode) {printf("Wall texture destroyed\n");}
     }
 
-    if (blnPole) {
-        SDL_DestroyTexture(pole);
-        blnPole = false;
+    if (blnloaded.blnPole) {
+        SDL_DestroyTexture(textures.pole);
+        blnloaded.blnPole = false;
         if (Global::blnDebugMode) {printf("Pole texture destroyed\n");}
     }
 
-    if (blnMonster) {
-        SDL_DestroyTexture(monster);
-        blnMonster = false;
+    if (blnloaded.blnMonster) {
+        SDL_DestroyTexture(textures.monster);
+        blnloaded.blnMonster = false;
         if (Global::blnDebugMode) {printf("Monster texture destroyed\n");}
     }
 
-    if (blnErrortex) {
-        SDL_DestroyTexture(errortex);
-        blnErrortex = false;
+    if (blnloaded.blnErrortex) {
+        SDL_DestroyTexture(textures.errortex);
+        blnloaded.blnErrortex = false;
         if (Global::blnDebugMode) {printf("Error texture destroyed\n");}
     }
 
-    if (blnMessageFont) {
+    if (blnloaded.blnMessageFont) {
         TTF_CloseFont(MessageFont);
-        blnMessageFont = false;
+        blnloaded.blnMessageFont = false;
         if (Global::blnDebugMode) {printf("Message font destroyed\n");}
     }
 
-    if (blnMessage) {
-        SDL_DestroyTexture(texmessage);
-        blnMessage = false;
+    if (blnloaded.blnMessage) {
+        SDL_DestroyTexture(textures.texmessage);
+        blnloaded.blnMessage = false;
         if (Global::blnDebugMode) {printf("Message texture destroyed\n");}
     }
 
-	if (blnRenderer) {
+	if (blnloaded.blnRenderer) {
         SDL_DestroyRenderer(ren);
-        blnRenderer = false;
+        blnloaded.blnRenderer = false;
         if (Global::blnDebugMode) {printf("Renderer destroyed\n");}
     }
-	if (blnWindow) {
+	if (blnloaded.blnWindow) {
         SDL_DestroyWindow(win);
-        blnWindow = false;
+        blnloaded.blnWindow = false;
         if (Global::blnDebugMode) {printf("Window destroyed\n");}
     }
 }
@@ -287,7 +285,7 @@ SDL_Texture* clsScreen::loadIMG(std::string filename) {
 
 	if (temp == nullptr) {
         printf("Failed to load img.\n");
-        return errortex;
+        return textures.errortex;
 	} else {
 	    if (Global::blnDebugMode) {printf("img to surface successful\n");}
     }
@@ -296,7 +294,7 @@ SDL_Texture* clsScreen::loadIMG(std::string filename) {
 	SDL_FreeSurface(temp);
 	if (tex == nullptr) {
         printf("Failed to create texture.\n");
-        tex = errortex;
+        tex = textures.errortex;
 	} else {
 	    if (Global::blnDebugMode) {printf("Surface to texture successful\n");}
     }
@@ -311,7 +309,7 @@ void clsScreen::playerDeath(void) {
 	//3 spaces then down about 4 spaces (depending on starting point)
 	//the whole thing happens in 5 frames.
 
-    /* TODO (GamerMan7799#1#): Add tileDeadPlayer with its own image to better till when the death animation is happening */
+    /* TODO (GamerMan7799#1#): Add tileDeadPlayer with its own image to better tell when the death animation is happening */
 	PLYR tempPlayer;
 	tempPlayer = Global::Enty.getPlayer();
 	//show();
@@ -333,23 +331,21 @@ void clsScreen::playerDeath(void) {
 /**********************************************************************************************************************************************************************/
 void clsScreen::writemessage(void) {
     //Now work on Making the messages that will appear on the screen
-    //Will require set up of Library SDL_TTF
-    //Which I don't want to do right now so comment it all out.
 
     //I imagine that I am doing terrible programing things
     //With how I'm switch between char's and strings but whatever
 
     //These are char arrays that will act as "Strings" for building the messages to appear
     char strClock[8];
-    char strGenNum[4];
-    char strPlayerNum[3];
-    char strFitness[6];
+    char strGenNum[5];
+    char strPlayerNum[4];
+    char strFitness[7];
 
     /* TODO (GamerMan7799#5#): Somehow detect if over wall and change color of text */
     std::string message;
     sprintf(strClock, "%8u", Global::Tick.getClockTime());
 
-    SDL_Surface* surmessage = TTF_RenderText_Solid(MessageFont, strClock, Black);
+    SDL_Surface* surmessage = TTF_RenderText_Solid(MessageFont, strClock, colors.Black);
     if (surmessage == nullptr) {
         printf("Failed to make clock surface.\n");
         error();
@@ -358,42 +354,51 @@ void clsScreen::writemessage(void) {
         if (Global::blnDebugMode) {printf("Surface Message successfully created\n");}
     }
 
-    texmessage = SDL_CreateTextureFromSurface(ren, surmessage);
-    if (texmessage == nullptr) {
+    textures.texmessage = SDL_CreateTextureFromSurface(ren, surmessage);
+    if (textures.texmessage == nullptr) {
         printf("Failed to convert message surface to texture.\n");
         error();
         return;
     } else {
         if (Global::blnDebugMode) {printf("Surface texture successfully created\n");}
-        blnMessage = true;
+        blnloaded.blnMessage = true;
     }
 
     SDL_Rect dst;
-    SDL_QueryTexture(texmessage,NULL,NULL, &dst.w, &dst.h);
+    SDL_QueryTexture(textures.texmessage,NULL,NULL, &dst.w, &dst.h);
     dst.y = 0;
     dst.x = width - dst.w;
-    SDL_RenderCopy(ren, texmessage, NULL, &dst);
+    SDL_RenderCopy(ren, textures.texmessage, NULL, &dst);
     if (Global::blnDebugMode) {printf("Clock placed.\n");}
 
     PLYR tempPlayer = Global::Enty.getPlayer();
 
     //Covert all the values into their char arrays
     int valuesscanned;
-    /* FIXME (GamerMan7799#1#): Program crashes if -Os -O3 or -s are enabled in compiler. */
+
     //Make the status string
     //Also make sure that each values has at least 1 values scanned in to avoid an error
     message = "Generation: ";
     valuesscanned = sprintf(strGenNum, "%2u", Global::Enty.uchrGenNum);
-    if (valuesscanned >= 1) {message += strGenNum;}
+    if (valuesscanned >= 1) {
+            strGenNum[4] = '\0'; //Make sure that the char is null terminated or the program crashes.
+            message += std::string(strGenNum);
+    }
     message += "     Player: ";
     valuesscanned = sprintf(strPlayerNum, "%3u", Global::Enty.uchrPlayerNum);
-    if (valuesscanned >= 1) {message += strPlayerNum;}
+    if (valuesscanned >= 1) {
+            strPlayerNum[3] = '\0'; //Make sure that the char is null terminated or the program crashes.
+            message += std::string(strPlayerNum);
+    }
     message += "     Fitness: ";
     valuesscanned = sprintf(strFitness, "%3.2f", tempPlayer.fitness);
-   if (valuesscanned >= 1) { message += strFitness;}
-    if (Global::blnDebugMode) {printf("Message written.\n");}
+    if (valuesscanned >= 1) {
+            strFitness[6] = '\0'; //Make sure that the char is null terminated or the program crashes.
+            message += std::string(strFitness);
+    }
+    if (Global::blnDebugMode) {printf("Status message made.\n");}
 
-    surmessage = TTF_RenderText_Solid(MessageFont, message.c_str(),White);
+    surmessage = TTF_RenderText_Solid(MessageFont, message.c_str(),colors.White);
     if (surmessage == nullptr) {
         printf("Failed to make status surface.\n");
         error();
@@ -401,22 +406,22 @@ void clsScreen::writemessage(void) {
     } else {
         if (Global::blnDebugMode) {printf("Generation Status Message successfully created.\n");}
     }
-    texmessage = SDL_CreateTextureFromSurface(ren, surmessage);
-    if (texmessage == nullptr) {
+    textures.texmessage = SDL_CreateTextureFromSurface(ren, surmessage);
+    if (textures.texmessage == nullptr) {
         printf("Failed to convert message surface to texture.\n");
-        blnMessage = false;
+        blnloaded.blnMessage = false;
         error();
         return;
     } else {
         if (Global::blnDebugMode) {printf("Surface texture successfully created\n");}
-        blnMessage = true;
+        blnloaded.blnMessage = true;
     }
 
-    SDL_QueryTexture(texmessage,NULL,NULL, &dst.w, &dst.h);
+    SDL_QueryTexture(textures.texmessage,NULL,NULL, &dst.w, &dst.h);
     dst.x = (int)(width - dst.w);
     dst.y = height - 30;
 
-    SDL_RenderCopy(ren, texmessage, NULL, &dst);
+    SDL_RenderCopy(ren, textures.texmessage, NULL, &dst);
     SDL_FreeSurface(surmessage);
 }
 /**********************************************************************************************************************************************************************/
@@ -441,5 +446,30 @@ SDL_Texture* clsScreen::loadERROR(void) {
     }
 
 	return tex;
+}
+/**********************************************************************************************************************************************************************/
+SDL_Rect clsScreen::detectEdge(SDL_Rect movingtex, SDL_Rect wall) {
+    //This will check if the moving tex is going to overlap the other texture.
+    //if it is, it will return a new rectangle ending of the edge.
+
+    if ( movingtex.x + movingtex.w >= wall.x && movingtex.x + movingtex.w <= wall.x + wall.w  ) {
+        //moving tex is above the wall x range
+        if (movingtex.y + movingtex.h >= wall.y && movingtex.y + movingtex.h <= wall.y + wall.h) {
+            //The lower edge is within the other texture box.
+            //Push moving texture up.
+            movingtex.y = wall.y - movingtex.h;
+            return movingtex;
+        } else if (movingtex.y >= wall.y && movingtex.y <= wall.y + wall.h) {
+            //The upper edge is in the box
+            //Push moving texture down
+            movingtex.y = wall.y + wall.h;
+            return movingtex;
+        } else {
+            //Not in the box
+            return movingtex;
+        }
+    } else {
+        return movingtex;
+    }
 }
 /**********************************************************************************************************************************************************************/